refactor : 클래스별 유효성 검증 코드 분리 및 테스트코드 수정
